add and move test
